# api-gateway/swagger/swagger.yaml
basePath: /
definitions:
...
  model.ErrorResponse:
    properties:
      code:
        example: 404
        type: integer
      message:
        example: resource not found
        type: string
    type: object
  model.GetAllUsersRequest:
    properties:
      page:
        example: 1
        minimum: 1
        type: integer
      page_size:
        example: 10
        maximum: 100
        minimum: 1
        type: integer
    required:
    - page
    - page_size
    type: object
...
  model.UpdateUserRequest:
    properties:
      email:
        example: user@example.com
        type: string
      username:
        example: newusername
        maxLength: 32
        minLength: 3
        type: string
    type: object
  model.User:
    properties:
      email:
        example: user@example.com
        type: string
      id:
        example: "1"
        type: string
      intersection_ids:
        example:
        - '[1'
        - "2"
        - 3]
        items:
          type: string
        type: array
      is_admin:
        example: false
        type: boolean
      username:
        example: johndoe
        type: string
    type: object
host: localhost:9090
info:
  contact:
    email: insideinsights2025@gmail.com
    name: Inside Insights Team
    url: https://swagger.io/
  description: This is the API Gateway for the Swift-Signals project, forwarding requests
    to backend gRPC microservices.
  termsOfService: http://example.com/terms/
  title: Authentication API Gateway
  version: "1.0"
paths:
  /admin/users:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of all users. Only accessible by admins.
      parameters:
      - description: Pagination options
        in: body
        name: getAllUsersRequest
        required: true
        schema:
          $ref: '#/definitions/model.GetAllUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "403":
          description: Forbidden - Only admins can access this endpoint
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Users
      tags:
      - Admin
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user by their ID. Only accessible by admins.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: User deleted successfully
        "403":
          description: Forbidden - Only admins can access this endpoint
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete User by ID
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Retrieves a user by their ID. Only accessible by admins.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/model.User'
        "403":
          description: Forbidden - Only admins can access this endpoint
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User by ID
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      description: Updates a user's details by their ID. Only accessible by admins.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated user details
        in: body
        name: updateUserRequest
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user details
          schema:
            $ref: '#/definitions/model.User'
        "403":
          description: Forbidden - Only admins can access this endpoint
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update User by ID
      tags:
      - Admin
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"